Move configuration to separate file
