+++ src/main/java/pl/cleankod/ApplicationInitializer.java
@@ -1,110 +1,14 @@
 package pl.cleankod;
 
-import feign.Feign;
-import feign.httpclient.ApacheHttpClient;
-import feign.jackson.JacksonDecoder;
-import feign.jackson.JacksonEncoder;
-import org.mapstruct.factory.Mappers;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.core.env.Environment;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import pl.cleankod.exchange.core.gateway.AccountRepository;
-import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
-import pl.cleankod.exchange.core.mapper.AccountMapper;
-import pl.cleankod.exchange.core.mapper.MoneyMapper;
-import pl.cleankod.exchange.core.service.AccountService;
-import pl.cleankod.exchange.core.service.AccountServiceImpl;
-import pl.cleankod.exchange.core.trace.ContextInitializeInterceptor;
-import pl.cleankod.exchange.core.trace.WebMvcConfig;
-import pl.cleankod.exchange.core.usecase.FindAccountAndConvertCurrencyUseCase;
-import pl.cleankod.exchange.core.usecase.FindAccountUseCase;
-import pl.cleankod.exchange.entrypoint.AccountController;
-import pl.cleankod.exchange.entrypoint.ExceptionHandlerAdvice;
-import pl.cleankod.exchange.provider.AccountInMemoryRepository;
-import pl.cleankod.exchange.provider.CurrencyConversionNbpService;
-import pl.cleankod.exchange.provider.nbp.ExchangeRatesNbpClient;
+import org.springframework.context.annotation.Import;
+import pl.cleankod.exchange.config.ApplicationConfiguration;
 
-import java.util.Currency;
-
-@SpringBootConfiguration
 @EnableAutoConfiguration
+@Import(ApplicationConfiguration.class)
 public class ApplicationInitializer {
     public static void main(String[] args) {
         SpringApplication.run(ApplicationInitializer.class, args);
-    }
-
-    @Bean
-    AccountRepository accountRepository() {
-        return new AccountInMemoryRepository();
-    }
-
-    @Bean
-    ExchangeRatesNbpClient exchangeRatesNbpClient(Environment environment) {
-        String nbpApiBaseUrl = environment.getRequiredProperty("provider.nbp-api.base-url");
-        return Feign.builder()
-                .client(new ApacheHttpClient())
-                .encoder(new JacksonEncoder())
-                .decoder(new JacksonDecoder())
-                .target(ExchangeRatesNbpClient.class, nbpApiBaseUrl);
-    }
-
-    @Bean
-    CurrencyConversionService currencyConversionService(ExchangeRatesNbpClient exchangeRatesNbpClient) {
-        return new CurrencyConversionNbpService(exchangeRatesNbpClient);
-    }
-
-    @Bean
-    FindAccountUseCase findAccountUseCase(AccountRepository accountRepository) {
-        return new FindAccountUseCase(accountRepository);
-    }
-
-    @Bean
-    FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase(
-            AccountRepository accountRepository,
-            CurrencyConversionService currencyConversionService,
-            Environment environment
-    ) {
-        Currency baseCurrency = Currency.getInstance(environment.getRequiredProperty("app.base-currency"));
-        return new FindAccountAndConvertCurrencyUseCase(accountRepository, currencyConversionService, baseCurrency);
-    }
-
-    @Bean
-    AccountController accountController(AccountService accountService) {
-        return new AccountController(accountService);
-    }
-
-    @Bean
-    ExceptionHandlerAdvice exceptionHandlerAdvice() {
-        return new ExceptionHandlerAdvice();
-    }
-
-    @Bean
-    AccountMapper accountMapper() {
-        return Mappers.getMapper( AccountMapper.class);
-    }
-
-    @Bean
-    MoneyMapper moneyMapper() {
-        return Mappers.getMapper( MoneyMapper.class);
-    }
-
-    @Bean
-    ContextInitializeInterceptor contextInitializeInterceptor() {
-        return new ContextInitializeInterceptor();
-    }
-
-    @Bean
-    WebMvcConfigurer webMvcConfig(ContextInitializeInterceptor contextInitializeInterceptor) {
-        return new WebMvcConfig(contextInitializeInterceptor);
-    }
-
-    @Bean
-    AccountService accountService(FindAccountAndConvertCurrencyUseCase findAccountAndConvertCurrencyUseCase,
-                                     FindAccountUseCase findAccountUseCase,
-                                     AccountMapper accountMapper) {
-        return new AccountServiceImpl(findAccountAndConvertCurrencyUseCase, findAccountUseCase, accountMapper);
     }
 }
